remove unused modifiable tests
