+++ .travis.yml
@@ -109,22 +109,6 @@ jobs:
     #   script: bundle exec parallel_split_test spec/plugin/backend_spec.rb --tag vimproc
     #   os: osx
 
-      # Although travis has less cores than 4, it's better for splitting as
-      # groups with even execution time cannot be precisely determined using files size
-    # - name: 'Acceptance tests of window output 1'
-      # script:
-        # - bundle exec parallel_test spec/plugin/window/syntax -n 2 --group-by filesize --type rspec
-        # - bundle exec parallel_test spec/plugin/window/modifiable/commandline/substitute_spec.rb -n 2 --group-by filesize --type rspec
-      # os: linux
-    # - name: 'Acceptance tests of window output 2'
-      # script: bundle exec parallel_split_test ./spec/plugin/window/modifiable_columnwise/within_single_context_spec.rb ./spec/plugin/compatibility/visual_multi_spec.rb
-      # os: linux
-    # - name: 'Acceptance tests of window output 3'
-      # script: bundle exec parallel_split_test ./spec/plugin/window/modifiable_columnwise/across_multiple_contexts_spec.rb
-      # os: linux
-    # - name: 'Acceptance tests of window output 4'
-      # script: bundle exec parallel_split_test ./spec/plugin/window/modifiable_linewise_spec.rb
-      # os: linux
     - name: 'Acceptance tests of #backend#system'
       script: bundle exec parallel_split_test spec/plugin/backend_spec.rb --tag system
       os: linux
