--- spec/support/helpers/modifiable/write.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Helpers::Modifiable::Write
-  def write_with_confirmation
-    editor.send_keys_separately ':write', :enter, 'y'
-  end
-end
